+++ includes/color.h
@@ -6,7 +6,7 @@
 /*   By: aditsch <aditsch@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/08 20:20:25 by aditsch           #+#    #+#             */
-/*   Updated: 2017/02/15 15:42:05 by aditsch          ###   ########.fr       */
+/*   Updated: 2017/02/15 19:40:18 by aditsch          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,12 @@
 
 typedef struct	s_color
 {
-	float		r;
-	float		g;
-	float		b;
+	double		r;
+	double		g;
+	double		b;
 }				t_color;
 
-t_color			ft_mult_color(t_color c, float r);
+t_color			ft_mult_color(t_color c, double r);
 t_color			ft_cross_color(t_color a, t_color b);
 t_color			ft_add_color(t_color a, t_color b);
 t_color			ft_unit_color(t_color c);
