+++ srcs/vector_op1.c
@@ -6,26 +6,26 @@
 /*   By: aditsch <aditsch@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/04 16:20:47 by aditsch           #+#    #+#             */
-/*   Updated: 2017/02/15 15:34:18 by aditsch          ###   ########.fr       */
+/*   Updated: 2017/02/15 20:06:55 by aditsch          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <math.h>
 #include "vector.h"
 
-float		ft_length(t_vector v)
+double		ft_length(t_vector v)
 {
 	return (sqrt(v.x * v.x + v.y * v.y + v.z * v.z));
 }
 
-float		ft_squared_length(t_vector v)
+double		ft_squared_length(t_vector v)
 {
 	return (v.x * v.x + v.y * v.y + v.z * v.z);
 }
 
 void		ft_normalize(t_vector *v)
 {
-	float	factor;
+	double	factor;
 
 	factor = 1. / sqrt(v->x * v->x + v->y * v->y + v->z * v->z);
 	v->x *= factor;
@@ -33,7 +33,7 @@ void		ft_normalize(t_vector *v)
 	v->z *= factor;
 }
 
-float		ft_dot(t_vector vec_1, t_vector vec_2)
+double		ft_dot(t_vector vec_1, t_vector vec_2)
 {
 	return (vec_1.x * vec_2.x + vec_1.y * vec_2.y + vec_1.z * vec_2.z);
 }
