+++ srcs/vector_op2.c
@@ -6,7 +6,7 @@
 /*   By: aditsch <aditsch@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/04 16:20:47 by aditsch           #+#    #+#             */
-/*   Updated: 2017/02/13 17:11:42 by aditsch          ###   ########.fr       */
+/*   Updated: 2017/02/15 20:06:57 by aditsch          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,12 +28,12 @@ t_vector	ft_sub(t_vector a, t_vector b)
 	return ((t_vector){a.x - b.x, a.y - b.y, a.z - b.z});
 }
 
-t_vector	ft_mult(t_vector v, float r)
+t_vector	ft_mult(t_vector v, double r)
 {
 	return ((t_vector){v.x * r, v.y * r, v.z * r});
 }
 
-t_vector	ft_div(t_vector v, float r)
+t_vector	ft_div(t_vector v, double r)
 {
 	return ((t_vector){v.x / r, v.y / r, v.z / r});
 }
